Replaced Function-Based Views with Class-Based Views

+++ watchlist_app/api/views_fn_based_views.py
@@ -1,20 +1,22 @@
-from django.http import request
+from gzip import READ
+from anyio import maybe_async
+from django.http import request, JsonResponse
+from watchlist_app.models import Movie
 
 from rest_framework import status
 from rest_framework.response import Response
-from rest_framework.views import APIView
-
+from rest_framework.decorators import api_view
 from watchlist_app.api import serializers
-from watchlist_app.models import Movie
 
-class MovieList(APIView):
+@api_view(['GET','POST'])
+def movies(request):
 
-    def get(self,request):
+    if request.method == 'GET':
         movies = Movie.objects.all()
         serializer = serializers.MovieSerializer(movies, many=True)
         return Response(serializer.data)
 
-    def post(self,request):
+    if request.method == 'POST':
         serializer = serializers.MovieSerializer(data=request.data)
         if serializer.is_valid():
             serializer.save()
@@ -22,32 +24,26 @@ class MovieList(APIView):
         else:
             return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-class MovieDetails(APIView):
+@api_view(['GET','PUT','DELETE'])
+def movie(request, pk):
 
-    def get(self,request,pk):
-        try:
-            movie = Movie.objects.get(pk=pk)
-        except Movie.DoesNotExist:
-            return Response({'error': 'Movie not found'}, status=status.HTTP_400_BAD_REQUEST)
+    try:
+        movie = Movie.objects.get(pk=pk)
+    except Movie.DoesNotExist:
+        return Response({'error': 'Movie not found'}, status=status.HTTP_400_BAD_REQUEST)
+
+    if request.method == 'GET':
         serializer = serializers.MovieSerializer(movie)
         return Response(serializer.data)
 
-    def put(self,request,pk):
-        try:
-            movie = Movie.objects.get(pk=pk)
-        except Movie.DoesNotExist:
-            return Response({'error': 'Movie not found'}, status=status.HTTP_400_BAD_REQUEST)
+    if request.method == 'PUT':
         serializer = serializers.MovieSerializer(movie, data=request.data)
         if serializer.is_valid():
             serializer.save()
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         else:
-            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)        
+            return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
-    def delete(self,request,pk):
-        try:
-            movie = Movie.objects.get(pk=pk)
-        except Movie.DoesNotExist:
-            return Response({'error': 'Movie not found'}, status=status.HTTP_400_BAD_REQUEST)
+    if request.method == 'DELETE':
         movie.delete()
         return Response(status=status.HTTP_204_NO_CONTENT)
